backend/requirements.txt
flask
flask-cors
flask-mysql
mysql-connector-python
PyJWT
requests
transformers
torch
torchvision
torchaudio
